+++ exonum/src/storage/key_set_index.rs
@@ -96,10 +96,10 @@ where
     /// let snapshot = db.snapshot();
     /// let name = "name";
     /// let prefix = vec![123];
-    /// let index: KeySetIndex<_, u8> = KeySetIndex::with_prefix(name, prefix, &snapshot);
+    /// let index: KeySetIndex<_, u8> = KeySetIndex::with_prefix(name, &prefix, &snapshot);
     /// # drop(index);
     /// ```
-    pub fn with_prefix<S: AsRef<str>>(name: S, prefix: Vec<u8>, view: T) -> Self {
+    pub fn with_prefix<S: AsRef<str>, P: StorageKey>(name: S, prefix: &P, view: T) -> Self {
         KeySetIndex {
             base: BaseIndex::with_prefix(name, prefix, IndexType::KeySet, view),
             _k: PhantomData,